+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/ExecutionProof.java
@@ -26,32 +26,9 @@ import tech.pegasys.teku.infrastructure.ssz.tree.TreeNode;
 public class ExecutionProof
     extends Container4<ExecutionProof, SszBytes32, SszUInt64, SszUInt64, SszVector<SszByte>> {
 
-  public static class ExecutionProofSchema
-      extends ContainerSchema4<
-          ExecutionProof, SszBytes32, SszUInt64, SszUInt64, SszVector<SszByte>> {
-    public ExecutionProofSchema() {
-      super(
-          "ExecutionProof",
-          namedSchema("block_hash", SszPrimitiveSchemas.BYTES32_SCHEMA),
-          namedSchema("subnet_id", SszPrimitiveSchemas.UINT64_SCHEMA),
-          namedSchema("version", SszPrimitiveSchemas.UINT64_SCHEMA),
-          namedSchema(
-              "proof_data",
-              SszVectorSchema.create(
-                  SszPrimitiveSchemas.BYTE_SCHEMA,
-                  65536)) // Assuming max size of proof_data is 65536 bytes
-          );
-    }
+  private static final ExecutionProofSchema SSZ_SCHEMA = new ExecutionProofSchema();
 
-    @Override
-    public ExecutionProof createFromBackingNode(final TreeNode node) {
-      return new ExecutionProof(this, node);
-    }
-  }
-
-  public static final ExecutionProofSchema SSZ_SCHEMA = new ExecutionProofSchema();
-
-  private ExecutionProof(final ExecutionProofSchema type, final TreeNode node) {
+  public ExecutionProof(final ExecutionProofSchema type, final TreeNode node) {
     super(type, node);
   }
